fix: Inconsistincies with mediation protocol
When switching from using Keys in the mediation protocol to using DIDs
in DIDComm V2, there were some inconsistencies within the API. The body
of the requests were changes but the API was not, even though it was
already a breaking change from 1.0 to 2.0. As part of my change, I have
renamed 'keylist' to 'recipient' to more-so simplify changes in the
future, should we switch from using DIDs to something else.

Also, the previous `keylist` response returned `keys`, rather than
`dids`, which was inconsistent with the other `keylist` API calls and
responses. This has been renamed to `dids`.

Language in other parts of the document has also been modified to
reflect these changes.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>

fix: Inconsistincies with mediation protocol (comebacks)

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>

fix: Change coordinate mediation version from 2.0 to 3.0

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>