ci(tests): Swap the get-changed-files repo
The original repo being used has been unmaintained for several years at
this point and is using a feature of github actions that is broken. This
change swaps it out for one that has commits on it from this year and
should (in theory) work better than the old action.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>